tools: add quiche-client and quiche-server
This adds a client and server application that
supports both HTTP/0.9 and HTTP/3 by default,
with preference towards the latter.

Other major differences compared to the example
binaries are the ability to multiplex requests
to different URLS, and for the client to log
response bodies to an indicated folder.

Common code is extracted into a library crate
but there is suprisingly little in common
between these two apps.