ignore frames on closed and collected streams
Due to the fact that we have already dropped the state of the stream, we
cannot know whether the frame is valid or not. Before we would assume
that this is an error, but in some cases (e.g. spurious retransmissions
due to loss of ACK frames) it's possible to receive a valid frame on a
dropped stream.

Fixes #350.